requirements.txt
google.generativeai

streamlit